Modifications to read and init

+++ src/sensors/can/C610.cpp
@@ -9,7 +9,7 @@ int C610::read(CAN_message_t& msg) {
     if (msg.id != m_base_id + m_id) return 0;
 
     // early return if the bus ID does not match
-    if (msg.bus - 1 != m_bus_id) return 0;
+    if ((uint32_t)(msg.bus - 1) != m_bus_id) return 0;
 
     // set m_input from msg
     memcpy(&m_input, &msg, sizeof(CAN_message_t));
